+++ lib/src/utils/builders.dart
@@ -14,22 +14,43 @@ import 'package:flutter_week_view/src/widgets/zoomable_header_widget.dart';
 /// Contains default builders and formatters.
 class DefaultBuilders {
   /// Formats a day in YYYY-MM-DD format, e.g., 2020-01-15.
-  static String defaultDateFormatter(int year, int month, int day) => '$year-${Utils.addLeadingZero(month)}-${Utils.addLeadingZero(day)}';
+  static String defaultDateFormatter(
+    int year,
+    int month,
+    int day,
+  ) =>
+      '$year-${Utils.addLeadingZero(month)}-${Utils.addLeadingZero(day)}';
 
   /// Formats a hour in 24-hour HH:MM format, e.g., 15:00.
-  static String defaultTimeFormatter(TimeOfDay time) => '${Utils.addLeadingZero(time.hour)}:${Utils.addLeadingZero(time.minute)}';
+  static String defaultTimeFormatter(
+    TimeOfDay time,
+  ) =>
+      '${Utils.addLeadingZero(time.hour)}:${Utils.addLeadingZero(time.minute)}';
 
   /// Allows to calculate a top offset according to the specified hour row height.
-  static double defaultTopOffsetCalculator(TimeOfDay time, {TimeOfDay minimumTime = TimeOfDayUtils.min, double hourRowHeight = 60}) {
+  static double defaultTopOffsetCalculator(
+    TimeOfDay time, {
+    TimeOfDay minimumTime = TimeOfDayUtils.min,
+    double hourRowHeight = 60,
+  }) {
     TimeOfDay relative = time.subtract(minimumTime.asDuration);
     return (relative.hour + (relative.minute / 60)) * hourRowHeight;
   }
 
   /// Builds a date according to a list.
-  static DateTime defaultDateCreator(List<DateTime> dates, int index) => dates[index];
+  static DateTime defaultDateCreator(
+    List<DateTime> dates,
+    int index,
+  ) =>
+      dates[index];
 
   /// Builds the current time indicator builder.
-  static Widget defaultCurrentTimeIndicatorBuilder(DayViewStyle dayViewStyle, TopOffsetCalculator topOffsetCalculator, double hourColumnWidth, bool isRtl) {
+  static Widget defaultCurrentTimeIndicatorBuilder(
+    DayViewStyle dayViewStyle,
+    TopOffsetCalculator topOffsetCalculator,
+    double hourColumnWidth,
+    bool isRtl,
+  ) {
     List<Widget> children = [
       if (dayViewStyle.currentTimeRuleHeight > 0 && dayViewStyle.currentTimeRuleColor != null)
         Expanded(
@@ -49,7 +70,7 @@ class DefaultBuilders {
         ),
     ];
 
-    final timeIndicatorHight = math.max(
+    final timeIndicatorHeight = math.max(
       dayViewStyle.currentTimeRuleHeight,
       dayViewStyle.currentTimeCircleRadius * 2,
     );
@@ -59,7 +80,7 @@ class DefaultBuilders {
     }
 
     return Positioned(
-      top: topOffsetCalculator(TimeOfDay.now()) - timeIndicatorHight / 2,
+      top: topOffsetCalculator(TimeOfDay.now()) - timeIndicatorHeight / 2,
       left: isRtl ? 0 : hourColumnWidth,
       right: isRtl ? hourColumnWidth : 0,
       child: Row(children: children),
@@ -67,19 +88,23 @@ class DefaultBuilders {
   }
 
   /// Builds the time displayed on the side border.
-  static Widget defaultHourColumnTimeBuilder(HourColumnStyle hourColumnStyle, TimeOfDay time) => Text(
+  static Widget defaultHourColumnTimeBuilder(
+    HourColumnStyle hourColumnStyle,
+    TimeOfDay time,
+  ) =>
+      Text(
         hourColumnStyle.timeFormatter(time),
         style: hourColumnStyle.textStyle,
       );
 
   /// Builds the time displayed on the side border.
-  static Widget defaultEventWidgetBuilder(
-    FlutterWeekViewEvent event,
+  static Widget defaultEventWidgetBuilder<E extends FlutterWeekViewEventMixin<E>>(
+    E event,
     double height,
     double width, {
     TimeFormatter? timeFormatter,
   }) =>
-      FlutterWeekViewEventWidget(
+      FlutterWeekViewEventWidget<E>(
         event: event,
         height: height,
         width: width,
@@ -87,8 +112,18 @@ class DefaultBuilders {
       );
 
   /// The default day view style builder.
-  static DayViewStyle defaultDayViewStyleBuilder(DateTime date) => DayViewStyle.fromDate(date: date);
+  static DayViewStyle defaultDayViewStyleBuilder(
+    DateTime date,
+  ) =>
+      DayViewStyle.fromDate(
+        date: date,
+      );
 
   /// The default day view style builder.
-  static DayBarStyle defaultDayBarStyleBuilder(DateTime date) => DayBarStyle.fromDate(date: date);
+  static DayBarStyle defaultDayBarStyleBuilder(
+    DateTime date,
+  ) =>
+      DayBarStyle.fromDate(
+        date: date,
+      );
 }