+++ lib/src/utils/utils.dart
@@ -14,7 +14,7 @@ class Utils {
 
   /// Checks whether the provided date is the same year, month and day than the target date.
   static bool sameDay(DateTime date, [DateTime? target]) {
-    target = target ?? dateTimeGetter.now();
+    target = target ?? DateTime.now();
     return target.year == date.year && target.month == date.month && target.day == date.day;
   }
 
@@ -27,20 +27,4 @@ class Utils {
 
     return words.getRange(0, words.length - 1).join(' ');
   }
-}
-
-/// A class that allows getting the current time, with the benefit that it can be mocked in tests.
-///
-/// Instead of using DateTime.now() directly, use DateTimeGetter.now() ([dateTimeGetter] is a global
-/// variable set below). Then, on tests, DateTimeGetter can be replaced with a mock that returns
-/// predictable values.
-class NowDateTimeGetter {
-  DateTime now() => DateTime.now();
-}
-
-NowDateTimeGetter dateTimeGetter = NowDateTimeGetter();
-
-/// This method is just for extra clarity the [dateTimeGetter] should only be modified on tests.
-void injectDateTimeGetterForTest(NowDateTimeGetter testDateTimeGetter) {
-  dateTimeGetter = testDateTimeGetter;
 }